Fix size_t to uint32_t conversion

+++ lib/libutee/tee_api_panic.c
@@ -44,7 +44,7 @@ static TEE_Result check_access(uint32_t flags, void *buf, size_t len)
 		return TEE_ERROR_SECURITY;
 
 	if (IS_ENABLED(CFG_TA_STRICT_ANNOTATION_CHECKS))
-		return TEE_CheckMemoryAccessRights(flags, buf, len);
+		return TEE_CheckMemoryAccessRights(flags, buf, (uint32_t)len);
 
 	return TEE_SUCCESS;
 }
